feat: move Recent Sales to /market and add Industry, Structures stubs
Splits the sales table out of the Characters page so each top-level nav
section owns a single concern. Adds reserved nav entries for Industry
and Structures with empty pages for future development.
